# docs/swagger.yaml
basePath: /v1
definitions:
...
  swagger.ConflictError:
    properties:
      error:
        example: record already exists
        type: string
    type: object
  swagger.ConflictResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ConflictError'
        type: array
    type: object
...
  swagger.ForbiddenError:
    properties:
      error:
        example: forbidden
        type: string
    type: object
  swagger.ForbiddenResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ForbiddenError'
        type: array
    type: object
...
  swagger.InternalServerError:
    properties:
      error:
        example: the server encountered a problem
        type: string
    type: object
  swagger.InternalServerErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.InternalServerError'
        type: array
    type: object
...
  swagger.ServiceUnavailableError:
    properties:
      error:
        example: database is unavailable
        type: string
    type: object
  swagger.ServiceUnavailableResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ServiceUnavailableError'
        type: array
    type: object
...
  swagger.TietoevrySymptomInput:
    properties:
      additional_data:
        example: '{"intensity": "moderate", "duration": "30min"}'
        type: string
      category:
        example: joint
        type: string
      comment:
        example: Pain started after morning run
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      date:
        example: "2024-01-15"
        type: string
      id:
        example: 2d4f6aee-b62c-408e-85e1-07bd78f383a7
        type: string
      original_id:
        example: ""
        type: string
      pain_index:
        example: 8
        type: integer
      raw_id:
        example: raw_symptom_123
        type: string
      recovered:
        example: true
        type: boolean
      severity:
        example: 7
        type: integer
      side:
        example: left
        type: string
      source:
        example: polar
        type: string
      symptom:
        example: knee pain
        type: string
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      user_id:
        example: 7cffe6e0-3f28-43b6-b511-d836d3a9f7b5
        type: string
    required:
    - created_at
    - date
    - id
    - severity
    - source
    - symptom
    - updated_at
    - user_id
    type: object
  swagger.TietoevrySymptomResponse:
    properties:
      symptoms:
        $ref: '#/definitions/swagger.TietoevrySymptomInput'
    type: object
  swagger.TietoevrySymptomsBulkInput:
    properties:
      symptoms:
        items:
          $ref: '#/definitions/swagger.TietoevrySymptomInput'
        type: array
    required:
    - symptoms
    type: object
  swagger.TietoevryTestResultInput:
    properties:
      comment:
        example: User reached fatigue quickly
        type: string
      created_at:
        example: "2024-07-23T08:10:00Z"
        type: string
      data:
        example: '{"vo2": 45.5, "lactate": 2.1}'
        type: string
      id:
        example: d19f2c63-fc5a-4aeb-8b90-5fc7df5d1c0c
        type: string
      name:
        example: Threshold Test 1
        type: string
      test_event_date:
        example: "2024-07-22"
        type: string
      test_event_id:
        example: 8abf0123-4567-89ab-cdef-1234567890ab
        type: string
      test_event_name:
        example: Summer Test Event
        type: string
      test_event_template_test_id:
        example: 4b9f2cde-8aaa-4e01-8103-8d1cabc712aa
        type: string
      test_event_template_test_limits:
        example: '{"max": 70, "min": 45}'
        type: string
      test_event_template_test_name:
        example: VO2 Max Lab Test
        type: string
      timestamp:
        example: "2024-07-23T08:00:00Z"
        type: string
      type_id:
        example: 23d9453b-8d43-46f5-bb27-1db29a88456a
        type: string
      type_name:
        example: Aerobic Threshold Test
        type: string
      type_result_type:
        example: float
        type: string
      type_type:
        example: threshold
        type: string
      updated_at:
        example: "2024-07-23T08:10:01Z"
        type: string
      user_id:
        example: 7cffe6e0-3f28-43b6-b511-d836d3a9f7b5
        type: string
    type: object
  swagger.TietoevryTestResultResponse:
    properties:
      test_results:
        $ref: '#/definitions/swagger.TietoevryTestResultInput'
    type: object
  swagger.TietoevryTestResultsBulkInput:
    properties:
      test_results:
        items:
          $ref: '#/definitions/swagger.TietoevryTestResultInput'
        type: array
    required:
    - test_results
    type: object
...
  swagger.UnauthorizedError:
    properties:
      error:
        example: unauthorized
        type: string
    type: object
  swagger.UnauthorizedResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.UnauthorizedError'
        type: array
    type: object
...
  swagger.ValidationError:
    properties:
      parameter name:
        example: full description of the error
        type: string
    type: object
  swagger.ValidationErrorResponse:
    properties:
      errors:
        items:
          $ref: '#/definitions/swagger.ValidationError'
        type: array
    type: object
info:
  contact: {}
  description: API for integrating, analyzing, and visualizing sports and health data
  termsOfService: https://csc.fi/en/security-privacy-data-policy-and-open-source-policy/privacy/
  title: KUHA REST API
paths:
...
  /tietoevry/symptoms:
    get:
      consumes:
      - application/json
      description: Get all symptoms for a specific user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.TietoevrySymptomResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get symptoms by user ID
      tags:
      - Tietoevry - Symptoms
    post:
      consumes:
      - application/json
      description: Insert multiple symptoms for user (idempotent)
      parameters:
      - description: Symptom data
        in: body
        name: symptoms
        required: true
        schema:
          $ref: '#/definitions/swagger.TietoevrySymptomsBulkInput'
      produces:
      - application/json
      responses:
        "201":
          description: Symptoms processed successfully (idempotent operation)
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swagger.ConflictResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Insert symptoms (bulk)
      tags:
      - Tietoevry - Symptoms
  /tietoevry/test-results:
    get:
      consumes:
      - application/json
      description: Get all test results for a specific user
      parameters:
      - description: User ID (UUID)
        in: query
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/swagger.TietoevryTestResultResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Get test results by user ID
      tags:
      - Tietoevry - Test_Results
    post:
      consumes:
      - application/json
      description: Insert multiple test results for users (idempotent)
      parameters:
      - description: Test result data
        in: body
        name: test_results
        required: true
        schema:
          $ref: '#/definitions/swagger.TietoevryTestResultsBulkInput'
      produces:
      - application/json
      responses:
        "201":
          description: Test results processed successfully
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/swagger.ValidationErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/swagger.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/swagger.ForbiddenResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/swagger.ConflictResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/swagger.InternalServerErrorResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/swagger.ServiceUnavailableResponse'
      security:
      - BearerAuth: []
      summary: Insert test results (bulk)
      tags:
      - Tietoevry - Test_Results
...
securityDefinitions:
  BearerAuth:
    description: 'Use format: Bearer your_JWT_here'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"